requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.0
openpyxl>=3.1.2
scikit-learn>=1.4.0
numpy>=1.26.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
PyPDF2>=3.0.0
python-docx>=1.0.0
openai>=1.0.0

